decided to do opengl for rendering and also turned positions into floats instead of ints

+++ classes/object.cpp
@@ -4,12 +4,12 @@
 
 #include "object.h"
 
-Object::Object(const uint texture_index=0,int position[2]=nullptr) {
+Object::Object(const uint texture_index=0,float position[2]=nullptr) {
     tex_index = texture_index;
     if (position != nullptr) {
         set_position(position);
     } else {
-        constexpr int pos[2] = {0,0};
+        constexpr float pos[2] = {0,0};
         set_position(pos);
     }
 }
@@ -18,7 +18,7 @@ void Object::set_texture(const int new_texture) {
     tex_index = new_texture;
 }
 
-void Object::set_position(const int new_pos[2]) {
+void Object::set_position(const float new_pos[2]) {
     position[0] = new_pos[0];
     position[1] = new_pos[1];
 }
